+++ resources/globals.py
@@ -15,8 +15,11 @@
 #
 
 import time
+import os, os.path
 
 JEEDOM_COM = ''
+JEEDOM_WEB = ''
+
 KNOWN_DEVICES = {}
 NOWPLAYING_DEVICES = {}
 GCAST_DEVICES = {}
@@ -48,6 +51,12 @@ DEFAULT_NODISPLAY = ""
 cycle_factor = 2
 cycle = 0.5
 cycle_main = 2
+
+tts_language = 'fr-FR'
+tts_engine = 'picotts'
+tts_cacheenabled = True
+tts_cachefolderweb = os.path.abspath(os.path.join(os.path.dirname(os.path.dirname(__file__)), 'tmp'))
+tts_cachefoldertmp = os.path.join('/tmp/jeedom/', 'googlecast_tts')
 
 log_level = "info"
 pidfile = '/tmp/googlecast.pid'