+++ resources/googlecast.py
@@ -19,6 +19,8 @@
 # pylint: disable=E
 
 import os,re
+import os.path
+import shutil
 import logging
 import sys
 import argparse
@@ -30,6 +32,9 @@ import traceback
 import _thread as thread
 import socket
 import resource
+from pydub import AudioSegment
+import hashlib
+from gtts import gTTS
 
 import globals
 
@@ -43,7 +48,7 @@ except ImportError:
 
 try:
     import pychromecast.pychromecast.controllers.dashcast as dashcast
-    import pychromecast.pychromecast.controllers.plex as plex
+    #import pychromecast.pychromecast.controllers.plex as plex
     import pychromecast.pychromecast.controllers.spotify as Spotify
 except ImportError:
     logging.error("ERROR: One or several pychromecast controllers are not loaded !")
@@ -52,6 +57,7 @@ except ImportError:
 
 try:
     import pychromecast.pychromecast.customcontrollers.youtube as youtube
+    import pychromecast.pychromecast.customcontrollers.plex as plex
 except ImportError:
     logging.error("ERROR: Custom controller not loaded !")
     logging.error(traceback.format_exc())
@@ -78,6 +84,7 @@ class JeedomChromeCast :
         if scan_mode == False :
             self.being_shutdown = False
             self.is_recovering = False
+            self.disable_notif = False
             self.customplayer = None
             self.customplayername = ""
             self.nowplaying_lastupdated = 0
@@ -110,6 +117,9 @@ class JeedomChromeCast :
         except Exception :
             return False
 
+    def getCurrentVolume(self):
+        return int(self.gcast.status.volume_level*100)
+
     def startNowPlaying(self):
         if self.now_playing == False and self.online == True :
             logging.debug("JEEDOMCHROMECAST------ Starting monitoring of " + self.uuid)
@@ -122,12 +132,13 @@ class JeedomChromeCast :
 
     def new_cast_status(self, new_status):
         #logging.debug("JEEDOMCHROMECAST------ Status " + str(new_status))
-        self._internal_refresh_status(False)
-        self.sendNowPlaying(force=True)
+        if not self.disable_notif :
+            self._internal_refresh_status(False)
+            self.sendNowPlaying(force=True)
 
     def new_media_status(self, new_mediastatus):
         #logging.debug("JEEDOMCHROMECAST------ New media status " + str(new_mediastatus))
-        if self.now_playing==True and new_mediastatus.player_state!="BUFFERING" :
+        if self.now_playing==True and new_mediastatus.player_state!="BUFFERING" and not self.disable_notif :
             self._internal_send_now_playing()
 
     def new_connection_status(self, new_status):
@@ -198,11 +209,16 @@ class JeedomChromeCast :
                     elif playername == 'youtube' :
                         player = youtube.YouTubeController()
                         self.gcast.register_handler(player)
+                        time.sleep(5)
                     elif playername == 'spotify' :
                         player = spotify.SpotifyController(token)
                         self.gcast.register_handler(player)
+                        time.sleep(2)
                     elif playername == 'plex' :
-                        player = plex.PlexController()
+                        player = plex.PlexController(self.gcast)
+                        player.register_status_listener(self)
+                        self.gcast.register_handler(player)
+                        time.sleep(2)
                     else :
                         player = self.gcast.media_controller
                     logging.debug("JEEDOMCHROMECAST------ Initiating player " + str(player.namespace))
@@ -217,12 +233,15 @@ class JeedomChromeCast :
                 except Exception :
                     player = None
                     pass
-            self.sendNowPlaying(force=True)
+            #else:
+            #    time.sleep(0.2)
+            if not self.disable_notif :
+                self.sendNowPlaying(force=True)
             return self.customplayer
         return None
 
     def resetPlayer(self) :
-        if self.customplayer is not None :
+        if self.customplayer is not None and self.customplayername != 'media' :
             if self.gcast.socket_client and self.customplayer.namespace in self.gcast.socket_client._handlers :
                 del self.gcast.socket_client._handlers[self.customplayer.namespace]
             self.customplayer = None
@@ -245,6 +264,7 @@ class JeedomChromeCast :
     def _internal_get_status(self):
         if self.gcast.status is not None and self.online == True:
             uuid = self.uuid
+            playStatus = self.gcast.media_controller.status
             status = {
                 "uuid" : uuid,
                 "friendly_name" : self.gcast.device.friendly_name,
@@ -256,6 +276,11 @@ class JeedomChromeCast :
                 "display_name" : self.gcast.status.display_name if self.gcast.status.display_name is not None else globals.DEFAULT_NODISPLAY,
                 "status_text" : self.gcast.status.status_text if self.gcast.status.status_text!="" else globals.DEFAULT_NOSTATUS,
                 "is_busy" : not self.gcast.is_idle,
+                "title" : playStatus.title,
+                "artist" : playStatus.artist,
+                'series_title': playStatus.series_title,
+                "stream_type" : playStatus.stream_type,
+                "player_state" : playStatus.player_state,
             }
             return status
         else :
@@ -264,6 +289,7 @@ class JeedomChromeCast :
                 "friendly_name" : "", "is_stand_by" :  False, "is_active_input" : False,
                 "display_name" : globals.DEFAULT_NODISPLAY, "status_text" : globals.DEFAULT_NOSTATUS,
                 "app_id" : "", "is_busy" : False,
+                "title" : "", "artist" : "", 'series_title': "", "stream_type" : "", "player_state" : "",
             }
 
 
@@ -271,19 +297,17 @@ class JeedomChromeCast :
         prev_status = self.previous_status
         if len(prev_status) != len(new_status) or len(prev_status)==2 or 'volume_level' not in new_status :
             return True
-        if prev_status['volume_level'] != new_status['volume_level'] :
+        if prev_status['status_text'] != new_status['status_text'] :
             return True
-        if prev_status['app_id'] != new_status['app_id'] :
+        if prev_status['volume_level'] != new_status['volume_level'] :
             return True
         if prev_status['is_busy'] != new_status['is_busy'] :
-            return True
-        if prev_status['volume_level'] != new_status['volume_level'] :
             return True
         if prev_status['volume_muted'] != new_status['volume_muted'] :
             return True
         if prev_status['is_stand_by'] != new_status['is_stand_by'] :
             return True
-        if prev_status['status_text'] != new_status['status_text'] :
+        if prev_status['app_id'] != new_status['app_id'] :
             return True
         return False
 
@@ -299,7 +323,6 @@ class JeedomChromeCast :
 
     def getStatus(self):
         return self._internal_get_status()
-
 
     def _internal_trigger_now_playing_update(self) :
         try :
@@ -389,132 +412,321 @@ class JeedomChromeCast :
 # main method to manage actions
 # -------------------------------
 def action_handler(message):
-    #name = message['command']['name']
     rootcmd = message['cmd']
-    uuid = message['command']['device']['uuid']
+    uuid = message['device']['uuid']
 
-    cmd = 'NONE'
-    if 'cmd' in message['command'] :
-        cmd = message['command']['cmd']
-    app = 'media'
-    if 'app' in message['command'] :
-        app = message['command']['app']
-    value = None
-    if 'value' in message['command'] :
-        value = message['command']['value']
-
-    needSendStatus = True
-    logging.debug("ACTION------ " + rootcmd + " - " + cmd + ' - ' + uuid + ' - ' + str(value)+ ' - ' + app)
     if uuid in globals.KNOWN_DEVICES and uuid in globals.GCAST_DEVICES and rootcmd == "action":
+        hascallback=False
+        callback=''
+        if 'callback' in message['device'] :
+            hascallback=message['device']['callback'] if True is not None else False
+            callback=message['device']['callback'] if message['device']['callback'] is not None else ''
 
-        jcast = globals.GCAST_DEVICES[uuid]
-        gcast = jcast.gcast
-        if cmd == 'refresh':
-            logging.debug("ACTION------Refresh action")
-        elif cmd == 'reboot':
-            logging.debug("ACTION------Reboot action")
-            gcast.reboot()
-            time.sleep(5)
-        elif cmd == 'volume_up':
-            logging.debug("ACTION------Volumme up action")
-            gcast.volume_up(value if value!=None else 0.1)
-        elif cmd == 'volume_down':
-            logging.debug("ACTION------Volume down action")
-            gcast.volume_down(value if value!=None else 0.1)
-        elif cmd == 'volume_set':
-            logging.debug("ACTION------Volume set action")
-            gcast.set_volume(int(value)/100)
-        elif cmd == 'play' and app == 'media':
-            logging.debug("ACTION------Play action")
-            gcast.media_controller.play()
-        elif cmd == 'stop' and app == 'media':
-            logging.debug("ACTION------Stop action")
-            gcast.media_controller.stop()
-        elif cmd == 'rewind':
-            logging.debug("ACTION------Rewind action")
-            gcast.media_controller.rewind()
-        elif cmd == 'skip':
-            logging.debug("ACTION------Skip action")
-            gcast.media_controller.skip()
-        elif cmd == 'seek':
-            logging.debug("ACTION------Seek action")
-            gcast.media_controller.seek(0 if value is None else value)
-        elif cmd == 'start_app':
-            logging.debug("ACTION------Stop action")
-            gcast.start_app('' if value is None else value)
-        elif cmd == 'quit_app':
-            logging.debug("ACTION------Stop action")
-            gcast.quit_app()
-        elif cmd == 'pause' and app == 'media':
-            logging.debug("ACTION------Stop action")
-            gcast.media_controller.pause()
-        elif cmd == 'mute_on':
-            logging.debug("ACTION------Mute on action")
-            gcast.set_volume_muted(True)
-        elif cmd == 'mute_off':
-            logging.debug("ACTION------Mute off action")
-            gcast.set_volume_muted(False)
-        elif app is not None:
-            logging.debug("ACTION------Playing action " + cmd + ' for application ' + app)
+        source=''
+        if 'source' in message['device'] :
+            source=message['device']['source']
+
+        commandlist = message['command']
+        if not isinstance(commandlist, list) :
+            commandlist = [commandlist]
+
+        for command in commandlist :
+            cmd = 'NONE'
+            if 'cmd' in command :
+                cmd = command['cmd']
+            app = 'media'
+            if 'app' in command :
+                app = command['app']
+            appid = ''
+            if 'appid' in command :
+                appid = command['appid']
+            value = None
+            if 'value' in command :
+                value = command['value']
+            sleep=0
+            if 'sleep' in command :
+                sleep = command['sleep']
+            vol = None
+            if 'vol' in command :
+                try:
+                    vol = int(command['vol'])
+                    if not (0 <= vol <= 100):
+                        vol = None
+                except:
+                    vol = None
+
+            # specific parameters for apps
+            quit_app_before=True
+            if 'quit_app_before' in command :
+                quit_app_before = True if command['quit_app_before'] else False
+            quit=False
+            if 'quit' in command :
+                quit = True if command['quit'] else False
+            force_register=False
+            if 'force_register' in command :
+                force_register = True if command['force_register'] else False
+            wait=2
+            if 'wait' in command :
+                wait = int(command['wait']) if command['wait'].isnumeric() else 2
+
+            if app == 'tts' :
+                cmd = 'tts'
+                app = 'media'
+
+            needSendStatus = True
+            fallbackMode = True
+            logging.debug("ACTION------ " + rootcmd + " - " + cmd + ' - ' + uuid + ' - ' + str(value)+ ' - ' + app)
+
+            jcast = globals.GCAST_DEVICES[uuid]
+            gcast = jcast.gcast
             try:
-                quit_app_before=True
-                if 'quit_app_before' in message['command'] :
-                    quit_app_before = True if message['command']['quit_app_before'] else False
-                force_register=False
-                if 'force_register' in message['command'] :
-                    force_register = True if message['command']['force_register'] else False
-                wait=2
-                if 'wait' in message['command'] :
-                    wait = message['command']['wait'] if message['command']['wait'].isnumeric() else 1
-
-                if app == 'web':    # app=web|cmd=load_url|value=https://news.google.com,True,5
+                if app == 'media' :    # app=media|cmd=play_media|value=http://bit.ly/2JzYtfX,video/mp4,Mon film
+                    possibleCmd = ['play_media']
+                    if cmd in possibleCmd :
+                        fallbackMode=False
+                        player = jcast.loadPlayer('media', { 'quitapp' : quit_app_before}, None)
+                        eval( 'player.' + cmd + '('+ gcast_prepareAppParam(value) +')' )
+                elif app == 'web':    # app=web|cmd=load_url|value=https://news.google.com,True,5
                     force_register=True
                     possibleCmd = ['load_url']
                     if cmd in possibleCmd :
+                        fallbackMode=False
                         player = jcast.loadPlayer(app, { 'quitapp' : quit_app_before}, None)
                         eval( 'player.' + cmd + '('+ gcast_prepareAppParam(value) +')' )
                 elif app == 'youtube':  # app=youtube|cmd=play_video|value=fra4QBLF3GU
-                    if gcast.device.cast_type == 'cast' :
-                        possibleCmd = ['play_video', 'start_new_session', 'add_to_queue', 'update_screen_id', 'clear_playlist', 'play', 'stop', 'pause']
-                        if cmd in possibleCmd :
+                    possibleCmd = ['play_video', 'start_new_session', 'add_to_queue', 'update_screen_id', 'clear_playlist', 'play', 'stop', 'pause']
+                    if cmd in possibleCmd :
+                        fallbackMode=False
+                        if gcast.device.cast_type == 'cast' :
                             player = jcast.loadPlayer(app, { 'quitapp' : quit_app_before, 'wait': wait}, None)
                             eval( 'player.' + cmd + '('+ gcast_prepareAppParam(value) +')' )
-                    else :
-                        logging.error("ACTION------ YouTube not availble on Chromecast Audio")
+                        else :
+                            logging.error("ACTION------ YouTube not availble on Chromecast Audio")
                 elif app == 'spotify':  # app=spotify|cmd=launch_app|token=XXXXXX
                     possibleCmd = ['launch_app']
                     if cmd in possibleCmd :
-                        if 'token' not in message['command'] :
+                        fallbackMode=False
+                        if 'token' not in command :
                             logging.error("ACTION------ Token missing for Spotify")
                         else :
-                            player = jcast.loadPlayer(app, { 'quitapp' : quit_app_before, 'wait': wait}, message['command']['token'])
+                            player = jcast.loadPlayer(app, { 'quitapp' : quit_app_before, 'wait': wait}, command['token'])
                             player.launch_app()
                 elif app == 'backdrop':  # also called backdrop
                     if gcast.device.cast_type == 'cast' :
+                        fallbackMode=False
                         gcast.start_app('E8C28D3C')
                     else :
                         logging.error("ACTION------ Backdrop not availble on Chromecast Audio")
                 elif app == 'plex':            # app=plex|cmd=pause
                     quit_app_before=False
-                    possibleCmd = ['play', 'stop', 'pause']
+                    possibleCmd = ['play_media','play', 'stop', 'pause', 'next', 'previous']
                     if cmd in possibleCmd :
+                        fallbackMode=False
                         player = jcast.loadPlayer(app, { 'quitapp' : quit_app_before, 'wait': wait}, None)
                         eval( 'player.' + cmd + '('+ gcast_prepareAppParam(value) +')' )
-                else : # media        # app=media|cmd=play_media|value=http://bit.ly/2JzYtfX,video/mp4,Mon film
-                    possibleCmd = ['play', 'stop', 'pause', 'play_media']
-                    player = jcast.loadPlayer('media', { 'quitapp' : quit_app_before}, None)
-                    eval( 'player.' + cmd + '('+ gcast_prepareAppParam(value) +')' )
+
+                if fallbackMode==False :
+                    logging.debug("ACTION------Playing action " + cmd + ' for application ' + app)
             except Exception as e:
-                logging.error("ACTION------Error while playing "+app+" : %s" % str(e))
+                logging.error("ACTION------Error while playing action " +cmd+ " on app " +app+" : %s" % str(e))
                 logging.debug(traceback.format_exc())
 
+            # low level google cast actions
+            if fallbackMode==True :
+                try:
+                    if cmd == 'refresh':
+                        logging.debug("ACTION------Refresh action")
+                        fallbackMode=False
+                    elif cmd == 'reboot':
+                        logging.debug("ACTION------Reboot action")
+                        gcast.reboot()
+                        time.sleep(5)
+                        fallbackMode=False
+                    elif cmd == 'volume_up':
+                        logging.debug("ACTION------Volumme up action")
+                        gcast.volume_up(value if value is not None else 0.1)
+                        fallbackMode=False
+                    elif cmd == 'volume_down':
+                        logging.debug("ACTION------Volume down action")
+                        gcast.volume_down(value if value is not None else 0.1)
+                        fallbackMode=False
+                    elif cmd == 'volume_set':
+                        logging.debug("ACTION------Volume set action")
+                        gcast.set_volume(int(value)/100)
+                        fallbackMode=False
+                    elif cmd == 'start_app':
+                        logging.debug("ACTION------Start app action")
+                        appToLaunch = '' if value is None else value
+                        if appid != '':
+                            appToLaunch=appid
+                        gcast.start_app(appToLaunch)
+                        fallbackMode=False
+                    elif cmd == 'quit_app':
+                        logging.debug("ACTION------Stop action")
+                        gcast.quit_app()
+                        jcast.resetPlayer()
+                        fallbackMode=False
+                    elif cmd == 'mute_on':
+                        logging.debug("ACTION------Mute on action")
+                        gcast.set_volume_muted(True)
+                        fallbackMode=False
+                    elif cmd == 'mute_off':
+                        logging.debug("ACTION------Mute off action")
+                        gcast.set_volume_muted(False)
+                        fallbackMode=False
+                    elif cmd == 'tts':
+                        logging.debug("ACTION------TTS action")
+                        lang = globals.tts_language
+                        if 'lang' in command :
+                            lang = command['lang']
+                        engine = globals.tts_engine
+                        if 'engine' in command :
+                            engine = command['engine']
+                        need_duration=False
+                        if vol is not None or quit==True :
+                            need_duration=True
+                        url,duration,mp3filename=get_tts_data(value, lang, engine, need_duration)
+                        curvol = jcast.getCurrentVolume()
+                        if vol is not None :
+                            gcast.set_volume(vol/100)
+                            time.sleep(0.1)
+                        thumb=globals.JEEDOM_WEB + '/plugins/googlecast/desktop/images/tts.png'
+                        jcast.disable_notif = True
+                        player = jcast.loadPlayer(app, { 'quitapp' : False, 'wait': wait}, None)
+                        #player.play_media(url, 'audio/mp3', 'TTS', thumb=thumb, stream_type="LIVE");
+                        player.play_media(url, 'audio/mp3', 'TTS', thumb=thumb, add_delay=0.1, stream_type="LIVE");
+                        player.block_until_active(timeout=1);
+                        jcast.disable_notif = False
+                        if vol is not None :
+                            time.sleep(duration+0.2)
+                            gcast.set_volume(curvol/100)
+                            vol = None
+                        if quit:
+                            gcast.quit_app()
+                        if globals.tts_cacheenabled==False :
+                            try:
+                                os.remove(mp3filename)
+                            except OSError:
+                                pass
+                        fallbackMode=False
+                except Exception as e:
+                    logging.error("ACTION------Error while playing action " +cmd+ " on low level commands : %s" % str(e))
+                    logging.debug(traceback.format_exc())
+
+            # media/application controler level Google Cast actions
+            if fallbackMode==True :
+                try:
+                    player=gcast.media_controller
+                    if cmd == 'play':
+                        logging.debug("ACTION------Play action")
+                        player.play()
+                        fallbackMode=False
+                    elif cmd == 'stop':
+                        logging.debug("ACTION------Stop action")
+                        player.stop()
+                        fallbackMode=False
+                    elif cmd == 'rewind':
+                        logging.debug("ACTION------Rewind action")
+                        player.rewind()
+                        fallbackMode=False
+                    elif cmd == 'skip':
+                        logging.debug("ACTION------Skip action")
+                        player.skip()
+                        fallbackMode=False
+                    elif cmd == 'seek':
+                        logging.debug("ACTION------Seek action")
+                        player.seek(0 if value is None else value)
+                        fallbackMode=False
+                    elif cmd == 'pause':
+                        logging.debug("ACTION------Stop action")
+                        player.pause()
+                        fallbackMode=False
+                    elif cmd == 'sleep':
+                        logging.debug("ACTION------Sleep")
+                        time.sleep(int(value))
+                        fallbackMode=False
+                except Exception as e:
+                    logging.error("ACTION------Error while playing action " +cmd+ " on default media controler : %s" % str(e))
+                    logging.debug(traceback.format_exc())
+
+            if vol is not None :
+                logging.debug("ACTION------SET VOLUME OPTION")
+                time.sleep(0.1)
+                gcast.set_volume(vol/100)
+
+            if fallbackMode==True :
+                logging.debug("ACTION------Action " + cmd + " not implemented !")
+
+            if sleep>0 :
+                time.sleep(sleep)
+
+            if needSendStatus :
+                globals.GCAST_DEVICES[uuid].sendDeviceStatus()
+
+        if hascallback :
+            callbackret=manage_callback(uuid, callback)
+            callbackmsg={'callback' : 1,'source' : source, 'uuid' : uuid, 'result': callbackret}
+            globals.JEEDOM_COM.send_change_immediate(callbackmsg);
+
+    else :
+        logging.error("ACTION------ Device not connected !")
+        return False
+
+    return True
+
+def manage_callback(uuid, callback_type):
+    # todo things for callback before returning value
+    return True
+
+def get_tts_data(text, language, engine, calcduration):
+    cachepath=globals.tts_cachefolderweb
+    # manage cache in ram memory
+    symlinkpath=globals.tts_cachefoldertmp
+    ttstext = (u''+text).encode('ascii').decode('utf-8')
+    try:
+        os.stat(symlinkpath)
+    except:
+        os.mkdir(symlinkpath)
+    try:
+        os.stat(cachepath)
+    except:
+        os.symlink(symlinkpath, cachepath)
+    try:
+        file = hashlib.md5(text.encode('utf-8')).hexdigest()
+        filenamemp3=os.path.join(cachepath,file+'.mp3')
+        if not os.path.isfile(filenamemp3):
+            if engine == 'picotts':
+                filename=os.path.join(cachepath,file+'.wav')
+                os.system('pico2wave -l '+language+' -w '+filename+ ' "' +ttstext+ '"')
+                speach = AudioSegment.from_wav(filename)
+                start_silence = AudioSegment.silent(duration=300)
+                speach = start_silence + speach
+                speach.export(filenamemp3, format="mp3", bitrate="128k", tags={'albumartist': 'Jeedom', 'title': 'TTS', 'artist':'Jeedom'}, parameters=["-ar", "44100","-vol", "200"])
+                duration_seconds = speach.duration_seconds
+                try:
+                    os.remove(filename)
+                except OSError:
+                    pass
+            elif engine == 'gtts':
+                language=language.split('-')[0]
+                tts = gTTS(text=ttstext, lang=language)
+                tts.save(filenamemp3)
+                speach = AudioSegment.from_mp3(filenamemp3)
+                start_silence = AudioSegment.silent(duration=300)
+                speach = start_silence + speach
+                speach.export(filenamemp3, format="mp3", bitrate="128k", tags={'albumartist': 'Jeedom', 'title': 'TTS', 'artist':'Jeedom'}, parameters=["-ar", "44100","-vol", "200"])
+                duration_seconds = speach.duration_seconds
         else:
-            logging.debug("ACTION------NOT IMPLEMENTED : " + cmd)
-
-        if needSendStatus :
-            globals.GCAST_DEVICES[uuid].sendDeviceStatus()
-    return
-
+            if calcduration == True:
+                speach = AudioSegment.from_mp3(filenamemp3)
+                duration_seconds = speach.duration_seconds
+            else:
+                duration_seconds=0
+        logging.debug("TTS------Sentence: '" +ttstext+ "' ("+engine+","+language+")")
+        urltoplay=globals.JEEDOM_WEB+'/plugins/googlecast/tmp/'+file+'.mp3'
+    except Exception as e:
+        logging.error("CMD-TTS------Exception while generating tts file : %s" % str(e))
+    return urltoplay, duration_seconds, filenamemp3
 
 def gcast_prepareAppParam(params):
     if params is None or params == '':
@@ -530,15 +742,15 @@ def gcast_prepareAppParam(params):
             p = s2[1].strip()
         if p.isnumeric() :
             ret = ret + ',' + prefix + p
-        elif p == 'True' or p == 'False' or p == 'None' :
-            ret = ret + ',' + prefix + p
+        elif p.lower() == 'true' or p.lower() == 'false' or p.lower() == 'none' :
+            ret = ret + ',' + prefix + p[0].upper()+p[1:]
         else :
             if p.startswith( "'" ) and p.endswith("'") :    # if starts already with simple quote
-                withoutQuote = p[1:-1]
+                withoutQuote = p[1:-1].lower()
                 if withoutQuote.isnumeric() :
                     ret = ret + ',' + prefix + withoutQuote
-                elif withoutQuote=='True' or withoutQuote=='False' or withoutQuote=='None' :
-                    ret = ret + ',' + prefix + withoutQuote
+                elif withoutQuote=='true' or withoutQuote=='false' or withoutQuote=='none' :
+                    ret = ret + ',' + prefix + withoutQuote[0].upper()+withoutQuote[1:]
                 else :
                     ret = ret + ',' + prefix + p
             else :
@@ -647,6 +859,13 @@ def read_socket(cycle):
                     uuid = message['device']['uuid']
                     if uuid in globals.GCAST_DEVICES :
                         globals.GCAST_DEVICES[uuid].sendDeviceStatus()
+                elif message['cmd'] == 'cleanttscache':
+                    logging.debug('SOCKET-READ------Clean TTS cache')
+                    try:
+                        if os.path.exists(globals.tts_cachefoldertmp):
+                            shutil.rmtree(globals.tts_cachefoldertmp)
+                    except:
+                        pass
                 elif message['cmd'] == 'refreshall':
                     logging.debug('SOCKET-READ------Attempt a refresh on all devices')
                     for uuid in globals.GCAST_DEVICES :
@@ -775,6 +994,7 @@ def scanner(name):
                         "is_stand_by" : False, "is_active_input" : False,
                         "display_name" : globals.DEFAULT_NODISPLAY, "status_text" : globals.DEFAULT_NOSTATUS,
                         "app_id" : "", "is_busy" : False,
+                        "title" : "", "artist" : "", 'series_title': "", "stream_type" : "", "player_state" : "",
                     }
                     #globals.JEEDOM_COM.add_changes('devices::'+known, globals.KNOWN_DEVICES[known])
                     globals.JEEDOM_COM.send_change_immediate_device(known, globals.KNOWN_DEVICES[known])
@@ -860,6 +1080,10 @@ parser.add_argument("--loglevel", help="Log Level for the daemon", type=str)
 parser.add_argument("--pidfile", help="PID filname", type=str)
 parser.add_argument("--callback", help="Callback url", type=str)
 parser.add_argument("--apikey", help="Jeedom API key", type=str)
+parser.add_argument("--ttsweb", help="Jeedom Web server (for TTS)", type=str)
+parser.add_argument("--ttslang", help="Default TTS language", type=str)
+parser.add_argument("--ttsengine", help="Default TTS engine", type=str)
+parser.add_argument("--ttscache", help="Use cache", type=str)
 parser.add_argument("--socketport", help="Socket Port", type=str)
 parser.add_argument("--sockethost", help="Socket Host", type=str)
 parser.add_argument("--daemonname", help="Daemon Name", type=str)
@@ -884,6 +1108,14 @@ if args.callback:
     globals.callback = args.callback
 if args.apikey:
     globals.apikey = args.apikey
+if args.ttsweb:
+    globals.JEEDOM_WEB = args.ttsweb
+if args.ttslang:
+    globals.tts_language = args.ttslang
+if args.ttsengine:
+    globals.tts_engine = args.ttsengine
+if args.ttscache:
+    globals.tts_cacheenabled = False if int(args.ttscache)==0 else True
 if args.cycle:
     globals.cycle = float(args.cycle)
 if args.cyclemain:
@@ -921,6 +1153,9 @@ logging.info('GLOBAL------Socket port : '+str(globals.socketport))
 logging.info('GLOBAL------Socket host : '+str(globals.sockethost))
 logging.info('GLOBAL------PID file : '+str(globals.pidfile))
 logging.info('GLOBAL------Apikey : '+str(globals.apikey))
+logging.info('GLOBAL------TTS Jeedom server : '+str(globals.JEEDOM_WEB))
+logging.info('GLOBAL------TTS default language : '+str(globals.tts_language))
+logging.info('GLOBAL------TTS default engine : '+str(globals.tts_engine))
 logging.info('GLOBAL------Callback : '+str(globals.callback))
 logging.info('GLOBAL------Event cycle : '+str(globals.cycle))
 logging.info('GLOBAL------Main cycle : '+str(globals.cycle_main))